backend/fastapi/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
attrs==24.1.0
bcrypt==3.2.0
black==21.4b2
blinker==1.4
Brlapi==0.8.3
cachetools==5.4.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
cloudpickle==3.0.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
cryptography==43.0.0
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.2.18
defer==1.0.6
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@82a57ce0b70057685962b352535147d9a8118578
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
document==1.0
duplicity==0.8.21
effdet==0.4.1
email_validator==2.2.0
environs==9.5.0
exceptiongroup==1.2.2
fastapi==0.111.1
fastapi-cli==0.0.4
fasteners==0.14.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.2
fvcore==0.1.5.post20221221
google-api-core==1.34.1
google-auth==2.32.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
hydra-core==1.3.2
idna==3.3
imageio==2.34.2
importlib-metadata==4.6.4
imutils==0.5.4
iopath==0.1.9
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-text-splitters==0.2.2
langsmith==0.1.96
language-selector==0.1
launchpadlib==1.10.16
layoutparser==0.3.4
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libhwp==0.2.0
lockfile==0.12.2
louis==3.20.0
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
mdurl==0.1.2
milvus-lite==2.4.8
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
omegaconf==2.3.0
openai==1.37.1
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.2
pexpect==4.8.0
Pillow==9.5.0
portalocker==2.10.1
protobuf==3.20.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycocotools==2.0.8
pycparser==2.22
pycups==2.0.1
pydantic==1.10.17
pydantic_core==2.20.1
pydot==3.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymilvus==2.4.4
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pypdfium2==4.30.0
pyRFC3339==1.1
pytesseract==0.3.10
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
regex==2024.5.15
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scipy==1.14.0
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.13.1
systemd-python==234
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2024.7.24
timm==1.0.7
toml==0.10.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typeddict==0.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.30.3
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.0.0
xdg==5
xkit==0.0.0
yacs==0.1.8
yarl==1.9.4
zipp==1.0.0
